Error when updating non-image attachemnt in admin

+++ wpthumb.crop-from-position.php
@@ -69,6 +69,9 @@ function wpthumb_media_form_crop_position( $fields, $post ) {
  */
 function wpthumb_media_form_crop_position_save( $post, $attachment ){
 
+	if ( ! isset( $attachment['wpthumb_crop_pos'] ) )
+		return;
+	
 	if ( $attachment['wpthumb_crop_pos'] == 'center,center' )
 		delete_post_meta( $post['ID'], 'wpthumb_crop_pos' );
